refactor(repository-matcher): update file structure + separation
